split PersonValidator from PeopleValidator
