+++ src/components/Section/Relationships/People/People.jsx
@@ -5,7 +5,8 @@ import validate from '../../../../validators'
 import { Field, Accordion, Svg } from '../../../Form'
 import { newGuid } from '../../../Form/ValidationElement'
 import Person from './Person'
-import PeopleValidator, { PersonValidator } from '../../../../validators/people'
+import PeopleValidator from '../../../../validators/people'
+import PersonValidator from '../../../../validators/person'
 import SubsectionElement from '../../SubsectionElement'
 import SummaryProgress from '../../History/SummaryProgress'
 import PeopleCounter from './PeopleCounter'
