+++ src/validators.js
@@ -49,7 +49,8 @@ import MaritalValidator from './validators/marital'
 import CohabitantsValidator, { CohabitantValidator } from './validators/cohabitant'
 import RelationshipsValidator from './validators/relationships'
 import DivorceValidator from './validators/divorce'
-import PeopleValidator, { PersonValidator } from './validators/people'
+import PeopleValidator from './validators/people'
+import PersonValidator from './validators/person'
 import ForeignContactsValidator, { ForeignNationalValidator } from './validators/foreigncontacts'
 import ForeignDirectActivityValidator from './validators/foreigndirectactivity'
 import ForeignDirectInterestValidator from './validators/foreigndirectinterest'
